fix(shortint): use correct value to shift msg in bivariate bps
For it to work, the bivariate bps assumed that
the degree of the ciphertext was message modulus - 1
so that the ctleft would get shifted by message modulus
and end up in the carry part.

Although it worked in most of our cases since
we were using smart ops which would, extract the carry
then set the degree to message modulus - 1.

Howver in the case of trivial ciphertext, the degree is set to
the value of the clear text (not a security issue) as the trivial is not
secure by design. But that meanse the assumption of degree ==
message_modulus - 1 did not hold true meaning computations we be
incorrect.