chore(tfhe): force compilation fail on non-optimized build
We've had a lot of complaints about performances that turned out
to be users not turning on optimization using `release` or other means.

As FHE is already slow on its own in optimized build, building non-optimized
version makes little to no sense.

This commit adds a way to check via the build.rs the optimization level
and if it is to low, force fail the compilation, with a long message
explaining why the compilation failed, and how to fix it.

This might be a bit intrusive, but but ints sort of intended as non-optimized
make non sense.

As everything happens in the build.rs, it wont affect user how build
tfhe-rs using another build system than cargo.