feat(integer): improve scalar_add
By using the fact that one input is scalar
we can speed up computation of scalar_add by packing the blocks.

As the main loop as a log2 complexity, it means we just do one less
iteration, but its enough to bring a 15% to 20% improvements.

In other words a 64bits scalar add is as fast as a 32bit add

Same idea is applicable to scalar_sub and overflowing variants