feat(ntt) Merge stimulus/driver key-handling path
Reduce the != between stimuli generation and driver to the strict
minimum.
In the Ntt-BR, move the normalization step from bsk to ntt_backward.
Enable to have bit-accurate value for stimulus.
Update reference to concrete-ntt to use RTL twiddles instead of the
default one.

Enable to have same behavior as in Hw (when hpu feature is enabled)