feat(hpu-sim) Hpu_8b tests are now working
Fix object lifetime and hook memory release request to Drop trait.
Indeed, explict memory release must be issued to the mockup simulator to
prevent allocation of new memory for key and ct on each iteration.
Since ofst register for key and bank are not yet support, this was
responsible of behavioral error on successive execution.

Fix DOp execution, error in Mac operands order and in delta computation
on HpuPbsParameters.