feat(hpu-simu) Add debug capability and fix behavioral issue
Add trace capability inside the instruction scheduler.

Fix scheduling issue observed with isc_depth & 16b/32b/64b IOp.
It take some time to figure it, since it wasn't due to schedler issue but to the postponed execution of operations.
Indeed, the introduction of Read early release in the simulation model prevent the behavioral execution to be postponed after the Retire stage.