feat(hpu) Implement new version of the Hpu Instruction Set
In depth rework of the DOp/IOp definition.
Dop structure was almost the same, however, the asm definition/parser
were rework with the following objectives:
* Template as first class citizen
* Support of Immediate template
* Direct parser and conversion between Asm/Hex
* Replace deku (and it's associated endianess limitation) by
  bitfield_struct and manual parsing

IOp structure was deeply rework. The aims was to introduce more
flexibiliy.
* Support various number of Destination
* Support various number of Sources
* Support various number of Immediat values
* Support of multiple bitwidth (Not implemented yet in the Fpga
  firmware)

Add lot of information in the IOp prototype that could be used at
runtime for proper error handling.