chore(ci): setup centralized workflow launcher
To ease external contributions, this implements a workflow responsible for triggering workflows that are currently related to pull_request event.
To ensure security, changes on some files are allowed only to Zama members. Since workflows are included in this file, this protects against secret leaks.

For now only AWS fast CPU tests is targeted as a proof of concept.